requirements.txt
# Core dependencies
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
databases[sqlite]==0.8.0

# RAG and ML
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.0.340
openai==0.28.0
google-generativeai==0.3.0
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.3
pillow==10.1.0
pytesseract==0.3.10

# Utilities
requests==2.31.0
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
tenacity==8.2.3
tqdm==4.66.1

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1